+++ pages/index.tsx
@@ -56,9 +56,13 @@ export default function Home() {
 										{equine.name}
 									</Typography>
 								</Grid>
-								<Grid item xs={5} md={4} lg={3}>
-									<Typography color="gray">{equine.yard.name}</Typography>
-								</Grid>
+								{equine.yard ? (
+									<Grid item xs={5} md={4} lg={3}>
+										<Typography color="gray">{equine.yard.name}</Typography>
+									</Grid>
+								) : (
+									<></>
+								)}
 							</Grid>
 							<IconButton>
 								<KeyboardArrowRightIcon fontSize="large" />
